pyproject.toml
[project]
name = "computacao-grafica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "pygame>=2.6.1",
    "pyopengl>=3.1.10",
]
